Split `MyTextEdit` into `SearchableTextEdit` and `SearchableBoldableTextEdit`
