+++ vocabsieve/global_events.py
@@ -1,0 +1,23 @@
+from PyQt5.QtCore import QObject, QTimer
+import functools
+
+@functools.lru_cache()
+class GlobalObject(QObject):
+    """
+    We need this to enable the textedit widget to communicate with the main window
+    """
+
+    def __init__(self):
+        super().__init__()
+        self._events = {}
+
+    def addEventListener(self, name, func):
+        if name not in self._events:
+            self._events[name] = [func]
+        else:
+            self._events[name].append(func)
+
+    def dispatchEvent(self, name):
+        functions = self._events.get(name, [])
+        for func in functions:
+            QTimer.singleShot(0, func)
