+++ vocabsieve/main.py
@@ -17,11 +17,12 @@ from .dictionary import *
 from .api import LanguageServer
 from .ext.reader import ReaderServer
 from .ext.importer import KindleImporter, KoreaderImporter
+from .global_events import GlobalObject
+from .ui.searchable_text_edit import SearchableTextEdit
+from .ui.searchable_boldable_text_edit import SearchableBoldableTextEdit
 from .text_manipulation import *
-from .settings import *
 import sys
 import importlib
-import functools
 import requests
 import platform
 import time
@@ -40,121 +41,6 @@ if platform.system() == "Darwin":
     MOD = "Cmd"
 else:
     MOD = "Ctrl"
-
-@functools.lru_cache()
-class GlobalObject(QObject):
-    """
-    We need this to enable the textedit widget to communicate with the main window
-    """
-
-    def __init__(self):
-        super().__init__()
-        self._events = {}
-
-    def addEventListener(self, name, func):
-        if name not in self._events:
-            self._events[name] = [func]
-        else:
-            self._events[name].append(func)
-
-    def dispatchEvent(self, name):
-        functions = self._events.get(name, [])
-        for func in functions:
-            QTimer.singleShot(0, func)
-
-class MyTextEdit(QTextEdit):
-
-    def __init__(self):
-        super(MyTextEdit, self).__init__()
-        self.currentCharFormatChanged.connect(self.handleCurrentCharFormatChanged)
-
-    @pyqtSlot()
-    def mouseDoubleClickEvent(self, e):
-        super().mouseDoubleClickEvent(e)
-        GlobalObject().dispatchEvent("double clicked")
-        self.textCursor().clearSelection()
-        self.original = ""
-
-    @property
-    def textBoldedByTags(self):
-        """
-        Current text content with bolded words defined by <b/>, irrespective of 
-        `settings.value("bold_style")`.
-        """
-        if settings.value("bold_style", type=int) == BoldStyle.FONTWEIGHT.value:
-            return markdown_boldings_to_bold_tag_boldings(
-                # toMarkdown() includes erroneous newlines
-                self.toMarkdown()[:-2])
-        elif settings.value("bold_style", type=int) == BoldStyle.BOLDCHAR.value:
-            return bold_char_boldings_to_bold_tag_boldings(self.toPlainText())[0]
-        else: return self.toPlainText()
-
-    def keyPressEvent(self, e):
-        super().keyPressEvent(e)
-
-        # Every time the user writes "_", check if we need to perform the 
-        # substitution "__{word}__" -> "<b>{word}</b>"
-        if settings.value("bold_style", type=int) == BoldStyle.FONTWEIGHT.value \
-            and e.text() == settings.value("bold_char"):
-            self.performBoldSubstitutions()
-
-    def performBoldSubstitutions(self):
-        bold_tags_substituted, subs_performed = \
-            bold_char_boldings_to_bold_tag_boldings(self.toPlainText())
-
-        if subs_performed:
-            def rebold_previously_bolded_words(string: str):
-                for should_bold in set(re.findall(r"\*\*(.+?)\*\*", self.toMarkdown())):
-                    string = re.sub(
-                        re.escape(should_bold), 
-                        lambda match: apply_bold_tags(match.group(0)), 
-                        string)
-                return string
-
-            bold_tags_substituted = rebold_previously_bolded_words(
-                bold_tags_substituted)
-
-            # Save cursor position
-            old_cursor_pos = self.textCursor().position()
-            # Set text
-            self.setText(bold_tags_substituted)
-
-            # Move back by the 4 characters removed when substituting 
-            # "__{word}__" => "<b>{word}</b>" (note that `cursor` does not 
-            # consider "<b/>" when calling `cursor.setPosition()`
-            cursor = self.textCursor()
-            cursor.setPosition(old_cursor_pos - 4 * subs_performed)
-            self.setTextCursor(cursor)
-
-    @pyqtSlot()
-    def handleCurrentCharFormatChanged(self):
-        """ 
-        If `settings.value("bold_style") == BoldStyle.FONTWEIGHT.value`, bolded characters are
-        added to the text editor. By default, the user cannot type in non-bold
-        font adjacent to these characters, so we reset the font weight to 
-        non-bold every time it changes.
-        """
-
-        def ensureNormalFontWeight():
-            cursor = self.textCursor()
-            fmt = cursor.charFormat()
-            if fmt.fontWeight() != QFont.Weight.Normal:
-                fmt.setFontWeight(QFont.Weight.Normal)
-                cursor.setCharFormat(fmt)
-                self.setTextCursor(cursor)
-        ensureNormalFontWeight()
-
-    @property
-    def unboldedText(self):
-        if settings.value("bold_style", type=int) == BoldStyle.FONTWEIGHT.value:
-            # `.toPlainText()` strips <b/> for us
-            return self.toPlainText()
-        elif settings.value("bold_style", type=int) == BoldStyle.BOLDCHAR.value:
-            # Remove bolds using bold_char
-            return remove_bold_char_boldings(self.toPlainText())
-        elif settings.value("bold_style", type=int) == "<disabled>":
-            # Text was never bolded in the first place
-            return self.toPlainText()
 
 class DictionaryWindow(QMainWindow):
     def __init__(self):
@@ -247,17 +133,17 @@ class DictionaryWindow(QMainWindow):
         self.namelabel = QLabel(
             "<h2 style=\"font-weight: normal;\">" + app_title(False) + "</h2>")
         self.menu = QMenuBar(self)
-        self.sentence = MyTextEdit()
+        self.sentence = SearchableBoldableTextEdit()
         self.sentence.setPlaceholderText(
             "Sentence copied to the clipboard will show up here.")
         self.sentence.setMinimumHeight(50)
         #self.sentence.setMaximumHeight(300)
         self.word = QLineEdit()
         self.word.setPlaceholderText("Word will appear here when looked up.")
-        self.definition = MyTextEdit()
+        self.definition = SearchableTextEdit()
         self.definition.setMinimumHeight(70)
         #self.definition.setMaximumHeight(1800)
-        self.definition2 = MyTextEdit()
+        self.definition2 = SearchableTextEdit()
         self.definition2.setMinimumHeight(70)
         #self.definition2.setMaximumHeight(1800)
         self.tags = QLineEdit()
@@ -1052,7 +938,7 @@ class DictionaryWindow(QMainWindow):
             )
         self.setImage(None)
 
-    def process_defi_anki(self, w: MyTextEdit, display_mode):
+    def process_defi_anki(self, w: SearchableTextEdit, display_mode):
         "Process definitions before sending to Anki"
         print("display mode is", display_mode)
         if display_mode in ["Raw", "Plaintext"]:
