+++ vocabsieve/ui/searchable_text_edit.py
@@ -1,0 +1,13 @@
+from PyQt5.QtWidgets import *
+from PyQt5.QtGui import *
+from PyQt5.QtCore import *
+from ..global_events import GlobalObject
+
+class SearchableTextEdit(QTextEdit):
+
+    @pyqtSlot()
+    def mouseDoubleClickEvent(self, e):
+        super().mouseDoubleClickEvent(e)
+        GlobalObject().dispatchEvent("double clicked")
+        self.textCursor().clearSelection()
+        self.original = ""
